Don't remove servicehealth files when mask is given.
If the ADAGUC_DATASET_MASK environment variable is set in
adaguc-server-updatedatasets.sh the servicehealth directory
should not be cleaned or we lose previously updated datasets.

+++ Docker/adaguc-server-updatedatasets.sh
@@ -25,8 +25,10 @@ if [[ $1 ]]; then
   done
 
 else
-  # remove all old service status files such that only active services are monitored
-  rm -f /servicehealth/*
+  if [[ ! "${ADAGUC_DATASET_MASK}" ]] ; then
+      # remove all old service status files such that only active services are monitored
+      rm -f /servicehealth/*
+  fi
   # Update all datasets
   for configfile in /data/adaguc-datasets/*xml ;do
     filename=/data/adaguc-datasets/"${configfile##*/}"
